Fixed return type to long.

+++ skipshuffle/src/main/java/com/dontbelievethebyte/skipshuffle/MediaPlayerBroadcastReceiver.java
@@ -7,10 +7,10 @@ import android.util.Log;
 
 public class MediaPlayerBroadcastReceiver extends BroadcastReceiver{
     private static final String TAG = "PLAYER_BROADCAST_RECEIVER";
-    private int playlistID;
+    private long playlistID;
     private int playlistPosition;
 
-    public int getPlaylistID() {
+    public long getPlaylistID() {
         return playlistID;
     }
     public int getPlaylistPosition(){
@@ -20,11 +20,11 @@ public class MediaPlayerBroadcastReceiver extends BroadcastReceiver{
     @Override
     public void onReceive(Context context, Intent intent) {
         if(intent.hasExtra(SkipShuflleMediaPlayerCommandsContract.PLAYLIST_ID)){
-            playlistID = intent.getIntExtra(SkipShuflleMediaPlayerCommandsContract.PLAYLIST_ID, 0);
-            Log.v(TAG, Integer.toString(playlistID));
+            playlistID = intent.getLongExtra(SkipShuflleMediaPlayerCommandsContract.PLAYLIST_ID, 0);
+            Log.v(TAG, Long.toString(playlistID));
         }
         if(intent.hasExtra(SkipShuflleMediaPlayerCommandsContract.PLAYLIST_CURSOR_POSITION)){
-            playlistID = intent.getIntExtra(SkipShuflleMediaPlayerCommandsContract.PLAYLIST_ID, 0);
+            playlistID = intent.getLongExtra(SkipShuflleMediaPlayerCommandsContract.PLAYLIST_ID, 0);
             Log.v(TAG, Integer.toString(playlistPosition));
         }
     }
